requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
pydantic==2.5.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.7
python-dotenv==1.0.0

# AI/ML Dependencies
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
chromadb==0.4.22
openai==1.12.0
tiktoken==0.5.2
pypdf2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2

# Neo4j
neo4j==5.15.0

# Additional utilities
aiofiles==23.2.1
httpx==0.25.2